Fix [Yanagishima] bytes is negative (#368)
* Fix [Yanagishima] bytes is negative

Closed #367


* Fix CI error and add test

+++ src/main/java/yanagishima/model/db/Query.java
@@ -42,7 +42,7 @@ public class Query {
   private Integer elapsedTimeMillis;
 
   @Column(name = "result_file_size")
-  private Integer resultFileSize;
+  private Long resultFileSize;
 
   @Column(name = "linenumber")
   private Integer linenumber;
